refactor(goals): neutralize goal lock service filename
